Move C4QueryObserver tests to QueryTest from ObserverTest
